Added benchmark container that calculates updates per sec and hijacks
