+++ benchmark/requirements.txt
@@ -1,0 +1,1 @@
+kombu==4.2.2.post1
