+++ benchmark/supervisor.d/services.conf
@@ -1,0 +1,127 @@
+[inet_http_server]
+port = 0.0.0.0:9001
+
+[eventlistener:listener]
+command=/usr/local/bin/python -c "import %(program_name)s; %(program_name)s.run()"
+directory=/root/core
+priority=999
+process_name=%(program_name)s
+numprocs=1
+events=PROCESS_STATE
+autorestart=true
+stderr_logfile=/dev/fd/2
+stderr_logfile_maxbytes=0
+
+[program:configuration]
+command=/usr/local/bin/python -c "import %(program_name)s; %(program_name)s.run()"
+process_name=%(program_name)s
+numprocs=1
+directory=/root/core
+umask=022
+autostart=true
+autorestart=unexpected
+startsecs=1
+startretries=3
+exitcodes=0,2
+stopsignal=INT
+stopwaitsecs=10
+stopasgroup=true
+killasgroup=true
+serverurl=AUTO
+stdout_logfile=/dev/fd/1
+stdout_logfile_maxbytes=0
+
+[program:observer]
+command=/usr/local/bin/python -c "import %(program_name)s; %(program_name)s.run()"
+process_name=%(program_name)s
+numprocs=1
+directory=/root/core
+umask=022
+autostart=true
+autorestart=unexpected
+startsecs=1
+startretries=3
+exitcodes=0,2
+stopsignal=INT
+stopwaitsecs=10
+stopasgroup=true
+killasgroup=true
+serverurl=AUTO
+stdout_logfile=/dev/fd/1
+stdout_logfile_maxbytes=0
+
+[program:clock]
+command=/usr/local/bin/python -c "import scheduler; scheduler.run()"
+process_name=%(program_name)s
+numprocs=1
+directory=/root/core
+umask=022
+autostart=true
+autorestart=unexpected
+startsecs=1
+startretries=3
+exitcodes=0,2
+stopsignal=INT
+stopwaitsecs=10
+stopasgroup=true
+killasgroup=true
+serverurl=AUTO
+stdout_logfile=/dev/fd/1
+stdout_logfile_maxbytes=0
+
+[program:database]
+command=/usr/local/bin/python -c "import %(program_name)s; %(program_name)s.run()"
+process_name=%(program_name)s
+numprocs=1
+directory=/root/core
+umask=022
+autostart=true
+autorestart=unexpected
+startsecs=1
+startretries=3
+exitcodes=0,2
+stopsignal=INT
+stopwaitsecs=10
+stopasgroup=true
+killasgroup=true
+serverurl=AUTO
+stdout_logfile=/dev/fd/1
+stdout_logfile_maxbytes=0
+
+[program:detection]
+command=/usr/local/bin/python -c "import %(program_name)s; %(program_name)s.run()"
+process_name=%(program_name)s
+numprocs=1
+directory=/root/core
+umask=022
+autostart=true
+autorestart=unexpected
+startsecs=1
+startretries=3
+exitcodes=0,2
+stopsignal=INT
+stopwaitsecs=10
+stopasgroup=true
+killasgroup=true
+serverurl=AUTO
+stdout_logfile=/dev/fd/1
+stdout_logfile_maxbytes=0
+
+[program:mitigation]
+command=/usr/local/bin/python -c "import %(program_name)s; %(program_name)s.run()"
+process_name=%(program_name)s
+numprocs=1
+directory=/root/core
+umask=022
+autostart=false
+autorestart=unexpected
+startsecs=1
+startretries=3
+exitcodes=0,2
+stopsignal=INT
+stopwaitsecs=10
+stopasgroup=true
+killasgroup=true
+serverurl=AUTO
+stdout_logfile=/dev/fd/1
+stdout_logfile_maxbytes=0
